fix(e2e): translate snake_case backend DTOs to camelCase bridge shape
aionui-backend emits snake_case JSON (full_path, relative_path, is_dir,
is_file, file_path, operation, new_path, content_type, created_at), but
workspace/preview E2E tests assert against the legacy camelCase IPC
contract. Rather than rewrite assertions, the bridge now translates at
its own boundary:

- split helpers/bridge.ts into bridge/{invoke,routes,mappers,index}.ts
  (each under 200 lines; helpers/ stays at 11 children)
- add per-route mapResponse for fs.dir, fs.list, fs.rename, fs.snapshot.compare
  and preview-history.{list,save,get-content}
- add mapBody for document.convert (filePath → file_path) and
  preview-history.* (target.contentType → content_type, snapshotId →
  snapshot_id via normalizePreviewTarget)

Verified: 17/17 runnable tests pass (workspace-files 6, workspace-snapshot 6,
preview-panel 5); 2 UI tests skip gracefully without a pre-existing conversation.
