+++ tests/e2e/helpers/bridge/index.ts
@@ -1,0 +1,1 @@
+export { invokeBridge } from './invoke';
